+++ TradingSystem/nemo_api.cpp
@@ -26,7 +26,7 @@ public:
 		//minute ms초 이후 구매가 되는 방식 (최소 1 ms)
 		if (minute <= 0) minute = 1;
 		Sleep(minute);
-		std::srand(std::time(NULL));
+		std::srand(static_cast<unsigned int>(std::time(NULL)));
 		int ret = std::rand() % 10 * 100 + 5000;
 		return ret;
 	}
